libsyn123: make MSVC more happy, mostly with explicit single precision work
git-svn-id: svn://scm.orgis.org/mpg123/trunk@5267 35dc7657-300d-0410-a2e5-dc2837fedb53

+++ src/libsyn123/libsyn123.c
@@ -397,7 +397,7 @@ static void wave_generator(syn123_handle *sh, int samples)
 	for(int i=0; i<samples; ++i)
 		sh->workbuf[1][i] = 1;
 	/* Add individual waves. */
-	for(int c=0; c<sh->wave_count; ++c)
+	for(size_t c=0; c<sh->wave_count; ++c)
 		wave_add_buffer( sh->workbuf[1], samples, sh->fmt.rate, sh->waves+c
 		,	sh->workbuf[0] );
 }
